[I2C] Use 8 bit for I2C address selection
This may reduce build size slightly.

+++ src/_P025_ADS1115.ino
@@ -53,7 +53,7 @@ boolean Plugin_025(uint8_t function, struct EventStruct *event, String& string)
     case PLUGIN_WEBFORM_SHOW_I2C_PARAMS:
     {
       #define ADS1115_I2C_OPTION 4
-      const int i2cAddressValues[] = { 0x48, 0x49, 0x4A, 0x4B };
+      const uint8_t i2cAddressValues[] = { 0x48, 0x49, 0x4A, 0x4B };
       if (function == PLUGIN_WEBFORM_SHOW_I2C_PARAMS) {
         addFormSelectorI2C(F("i2c_addr"), ADS1115_I2C_OPTION, i2cAddressValues, PCONFIG(0));
       } else {
