+++ src/_P030_BMP280.ino
@@ -105,7 +105,7 @@ boolean Plugin_030(uint8_t function, struct EventStruct *event, String& string)
     case PLUGIN_I2C_HAS_ADDRESS:
     case PLUGIN_WEBFORM_SHOW_I2C_PARAMS:
     {
-      const int i2cAddressValues[] = { 0x76, 0x77 };
+      const uint8_t i2cAddressValues[] = { 0x76, 0x77 };
       if (function == PLUGIN_WEBFORM_SHOW_I2C_PARAMS) {
         addFormSelectorI2C(F("i2c_addr"), 2, i2cAddressValues, PCONFIG(0));
         addFormNote(F("SDO Low=0x76, High=0x77"));
